Small wording change
Word choice is difficult, so let this commit message be a footnote.
I'm thankful for the trans people in my life and I'm glad they're around.

It's hard to say "I'm happy you're here" in a way that doesn't seem
patronising, and I'm honestly not certain that this message is helpful.
I've guessed that it might be for a few reasons.

This change aims to revise away some more patronising interpretations of
what I'm trying to say here, at any rate.