Minor variable/symbol renaming
Some occam example code from INMOS gives boolean variables and
parameters names like active.not.inactive or automatic.not.manual. The
idea is that the name tells you what the variable means when it's true
(the left side of .not.) and when it's false (the right side of .not.).
This is helpful, but to the casual reader it could also seem like you're
intentionally dichotomising two concepts that aren't incompatible. This
renaming aims to avoid that impression and changes some related names
for consistency.